+++ vscode/webviews/promptEditor/BaseEditor.tsx
@@ -20,12 +20,10 @@ import { RICH_EDITOR_NODES } from './nodes'
 import MentionsPlugin from './plugins/atMentions/atMentions'
 import CodeHighlightPlugin from './plugins/codeHighlight'
 import { KeyboardEventPlugin, type KeyboardEventPluginProps } from './plugins/keyboardEvent'
-import { OnFocusPlugin } from './plugins/onFocus'
 
 interface Props extends KeyboardEventPluginProps {
     initialEditorState: SerializedEditorState | null
     onChange: (editorState: EditorState, editor: LexicalEditor) => void
-    onFocus?: () => void
     editorRef?: RefObject<LexicalEditor>
     placeholder?: string
     disabled?: boolean
@@ -39,7 +37,6 @@ interface Props extends KeyboardEventPluginProps {
 export const BaseEditor: FunctionComponent<Props> = ({
     initialEditorState,
     onChange,
-    onFocus,
     editorRef,
     placeholder,
     disabled,
@@ -87,7 +84,6 @@ export const BaseEditor: FunctionComponent<Props> = ({
                     <MentionsPlugin />
                     <CodeHighlightPlugin />
                     <AutoFocusPlugin />
-                    <OnFocusPlugin onFocus={onFocus} />
                     {editorRef && <EditorRefPlugin editorRef={editorRef} />}
                     <KeyboardEventPlugin
                         onKeyDown={onKeyDown}
